+++ include/imports.h
@@ -4,18 +4,16 @@
 //funcionalidades do programa
 #include <stdio.h>
 #include "database.h"
-#include "callbacks.h"
+#include "visual.h"
 #include "professores.h"
 #include "turmas.h"
 #include "alunos.h"
 #include "matriculas.h"
 #include "menu.h"
 
-#include <unistd.h> //para o sleep() funcionar
-//função de limpar a tela
+#include <string.h>
+#include <unistd.h> 
 #include <stdlib.h>
-
-//função tratar siginit
 #include <signal.h> 
 
 #endif
