+++ include/visual.h
@@ -1,5 +1,11 @@
-#ifndef CALLBACKS_H
-#define CALLBACKS_H
+#ifndef VISUAL_H
+#define VISUAL_H
+
+//limpar terminal
+void limpar_tela();
+
+//função de segurança, executa caso o programa seja interrompido
+void trata_sigint(int sig);
 
 //Callback
 int callback_alunos(void *data, int argc, char **argv, char **colName);
